Replace DrawImage implementation for draw.Scaler.
This change has no API modifications.
