+++ draw2dimg/ftgc.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"image"
 	"image/color"
-	"image/draw"
 	"log"
 	"math"
 
@@ -17,6 +16,7 @@ import (
 	"github.com/golang/freetype/raster"
 	"github.com/golang/freetype/truetype"
 
+	"golang.org/x/image/draw"
 	"golang.org/x/image/font"
 	"golang.org/x/image/math/fixed"
 )
@@ -37,6 +37,18 @@ type GraphicContext struct {
 	glyphBuf         *truetype.GlyphBuf
 	DPI              int
 }
+
+// ImageFilter defines the type of filter to use
+type ImageFilter int
+
+const (
+	// LinearFilter defines a linear filter
+	LinearFilter ImageFilter = iota
+	// BilinearFilter defines a bilinear filter
+	BilinearFilter
+	// BicubicFilter defines a bicubic filter
+	BicubicFilter
+)
 
 // NewGraphicContext creates a new Graphic context from an image.
 func NewGraphicContext(img draw.Image) *GraphicContext {
@@ -82,6 +94,25 @@ func (gc *GraphicContext) Clear() {
 func (gc *GraphicContext) ClearRect(x1, y1, x2, y2 int) {
 	imageColor := image.NewUniform(gc.Current.FillColor)
 	draw.Draw(gc.img, image.Rect(x1, y1, x2, y2), imageColor, image.ZP, draw.Over)
+}
+
+// DrawImage draws an image into dest using an affine transformation matrix, an op and a filter
+func DrawImage(src image.Image, dest draw.Image, tr draw2d.Matrix, op draw.Op, filter ImageFilter) {
+	srcRect := src.Bounds()
+	x0, y0, x1, y1 := tr.TransformRectangle(float64(srcRect.Min.X), float64(srcRect.Min.Y), float64(srcRect.Max.X), float64(srcRect.Max.Y))
+	destRect := image.Rectangle{image.Point{int(x0), int(y0)}, image.Point{int(x1), int(y1)}}
+
+	var scaler draw.Scaler
+	switch filter {
+	case LinearFilter:
+		scaler = draw.NearestNeighbor
+	case BilinearFilter:
+		scaler = draw.BiLinear
+	case BicubicFilter:
+		scaler = draw.CatmullRom
+	}
+
+	scaler.Scale(dest, destRect, src, srcRect, draw.Over, nil)
 }
 
 // DrawImage draws the raster image in the current canvas
